+++ src/main/java/com/backend/elearning/domain/lecture/LecturePostVM.java
@@ -3,7 +3,7 @@ package com.backend.elearning.domain.lecture;
 public record LecturePostVM(
         Long id,
         String title,
-        int number,
+        float number,
         String lectureDetails,
         String videoId,
         Long sectionId
